Fix the type of m_lastDiff
- Was an int by mistake

+++ src/game/ai.hpp
@@ -56,7 +56,7 @@ private:
 
     const Route * m_route;
 
-    int m_lastDiff;
+    float m_lastDiff;
 
     int m_lastTargetNodeIndex;
